samv7/pwm: add support for latched fault inputs and clear option
This adds configuration option for every fault input, that can be
latched (kept even after the input value is below the threshold) or
volatile (PWM is automatically enabled once the output polarity goes
back to the normal state).

The commit also adds the support for PWMIOC_FAULTS_CLEAR ioctl
that allows the user to clear latched fault inputs from the
application.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>